Adding python libs to source control
Only reason for not doing this was to keep repo small, that argument was
pretty stupid as you still had to download all the dependencies
afterwards. Since this repo won't be used as a component in another one
the benefits of having the deps in source outweigh the disadvantages
(suppose pypy goes down - suddenly I wouldn't be able to deploy code
changes). Deployments are much quicker with libs in source.